+++ spec/DefinitionSpec.php
@@ -13,28 +13,23 @@ namespace spec\Xabbuh\XApi\Model;
 
 use PhpSpec\ObjectBehavior;
 use Xabbuh\XApi\Model\Definition;
+use Xabbuh\XApi\Model\LanguageMap;
 
 class DefinitionSpec extends ObjectBehavior
 {
     function its_properties_can_be_read()
     {
+        $name = LanguageMap::create(array('en-US' => 'test'));
+        $description = LanguageMap::create(array('en-US' => 'test'));
         $this->beConstructedWith(
-            array('en-US' => 'test'),
-            array('en-US' => 'test'),
+            $name,
+            $description,
             'http://id.tincanapi.com/activitytype/unit-test',
             'https://github.com/adlnet/xAPI_LRS_Test'
         );
 
-        $name = $this->getName();
-        $name->shouldBeArray();
-        $name->shouldHaveCount(1);
-        $name->shouldHaveKeyWithValue('en-US', 'test');
-
-        $description = $this->getName();
-        $description->shouldBeArray();
-        $description->shouldHaveCount(1);
-        $description->shouldHaveKeyWithValue('en-US', 'test');
-
+        $this->getName()->shouldReturn($name);
+        $this->getDescription()->shouldReturn($description);
         $this->getType()->shouldReturn('http://id.tincanapi.com/activitytype/unit-test');
         $this->getMoreInfo()->shouldReturn('https://github.com/adlnet/xAPI_LRS_Test');
     }
@@ -51,11 +46,11 @@ class DefinitionSpec extends ObjectBehavior
 
     function it_is_different_when_names_are_omitted_and_other_definition_contains_an_empty_list_of_names()
     {
-        $this->equals(new Definition(array()))->shouldReturn(false);
+        $this->equals(new Definition(new LanguageMap()))->shouldReturn(false);
     }
 
     function it_is_different_when_descriptions_are_omitted_and_other_definition_contains_an_empty_list_of_descriptions()
     {
-        $this->equals(new Definition(null, array()))->shouldReturn(false);
+        $this->equals(new Definition(null, new LanguageMap()))->shouldReturn(false);
     }
 }